// src/app/model/producto.ts
export class producto {
    idproduct: number;
    dateadd: string;
    datemod: string;
    existence: number;
    quantity: number;
    iduseradd: number;
    idusermod: number;
    productname: string;
    purchaseprice: number;
    saleprice: number;
    unitaverage: string;
    totalprice: number;
}