// src/app/model/venta.ts
export class venta {
    idventa: number;
    bill: string;
    idproducto: number;
    productname: string;
    quantity: number;
    purchaseunitprice: number;
    totalpurchaseprice: number;
    totalunitsaleprice: number;
    totalsaleprice: number;
    ganancia: number;
    unitaverage: string;
    iduseradd: number;
    dateadd!: string;
    idusermod: number;
    datemod: string;
}