// src/app/model/pedido.ts
export class pedido {
    idpendiente: number;
    bill: string;
    nombredeudor: string;
    idproducto: number;
    productname: string;
    quantity: number;
    purchaseunitprice: number;
    totalpurchaseprice: number;
    totalunitsaleprice: number;
    totalsaleprice: number;
    ganancia: number;
    unitaverage: string;
    iduseradd: number;
    dateadd!: string;
    idusermod: number;
    datemod: string;
}